pyproject.toml
[project]
name = "yolo-draw"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "pillow",
    "pyside6",
    "pyinstaller",
    "ultralytics"
]